requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
spacy==3.7.4
nltk==3.9.1
streamlit==1.37.1
plotly==5.23.0
folium==0.17.0
streamlit-folium==0.20.0
geopy==2.4.1
PyMuPDF==1.24.10
gspread==6.1.2
oauth2client==4.1.3
wordcloud==1.9.3
scikit-learn==1.5.1

openai>=1.35.0
